apps/api/pyproject.toml
[project]
name = "api-app"
version = "0.1.0"
description = "FastAPI backend for Sernia Capital portfolio"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI and server
    "fastapi>=0.115.12",
    "fastapi-cli>=0.0.4",
    "hypercorn==0.17.3",
    "starlette>=0.40.0",
    "python-multipart>=0.0.18",

    # Scheduler (to be replaced by Prefect)
    "apscheduler>=3.11.0",

    # Database
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.10",
    "sqlalchemy[asyncio]>=2.0.41",

    # Auth
    "clerk-backend-api==1.7.2",
    "svix>=1.81.0",

    # AI
    "openai>=1.77.0",
    "pydantic-ai>=1.18.0",

    # GraphQL
    "strawberry-graphql[fastapi]>=0.219.2",

    # Google APIs
    "google-api-python-client>=2.108.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",

    # Twilio/SMS
    "twilio>=9.5.2",

    # Observability
    "logfire[asyncpg,httpx,fastapi,sqlalchemy,requests]>=0.11.0",

    # Core utilities
    "pydantic>=2.11.2",
    "python-dotenv>=1.0.1",
    "httpx==0.28.1",
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.4",
    "pytz>=2025.2",
    "email-validator==2.2.0",
    "jinja2==3.1.4",
    "typer==0.12.3",
    "itsdangerous>=2.2.0",
    "anyio>=4.8.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio>=0.25.2",
]
